Use correct poll interval
EurekaConnection.Pollinterval is already a time.Duration, so there's no
need to cast and multiple it by time.Second. This multiplication increases the
interval by a factor of 1000. So if the caller meant to set the interval to 1
second, the code will actually wait for 1000 seconds for each poll.

It looks like this bug was introduced by https://github.com/hudl/fargo/commit/0331c1f